More pure components. Delete unnecessary handlebars files
