+++ SingularityUI/app/actions/ui/starred.es6
@@ -1,9 +1,3 @@
-export const GET_STARRED_REQUESTS = 'GET_STARRED_REQUESTS';
-
-export const getStarredRequests = () => ({
-  type: GET_STARRED_REQUESTS
-});
-
 export const CHANGE_REQUEST_STAR = 'CHANGE_REQUEST_STAR';
 
 export const changeRequestStar = (requestId) => ({
