--- SingularityUI/app/templates/dashboard.hbs
@@ -1,113 +1,0 @@
-<header>
-    {{#if deployUser}}
-        <h1>{{ deployUser }}</h1>
-    {{else}}
-        <h1>Singularity</h1>
-    {{/if}}
-</header>
-
-{{#if collectionSynced}}
-    {{#if deployUser}}
-        <div class="row">
-            <div class="col-md-12">
-                <div class="page-header">
-                    <h2>My requests</h2>
-                </div>
-                <div class="row">
-                    <div class="col-md-2">
-                        <a class="big-number-link" href="{{appRoot}}/requests/active/all/{{ deployUser }}" data-requests-active-filter="all">
-                            <div class="well">
-                                <div class="big-number">
-                                    <div class="number" data-state-attribute="requests">
-                                        {{ userRequestTotals.all }}
-                                    </div>
-                                    <div class="number-label">total</div>
-                                </div>
-                            </div>
-                        </a>
-                    </div>
-                    <div class="col-md-2">
-                        <a class="big-number-link" href="{{appRoot}}/requests/active/ON_DEMAND/{{ deployUser }}" data-requests-active-filter="ON_DEMAND">
-                            <div class="well">
-                                <div class="big-number">
-                                    <div class="number" data-state-attribute="requests">
-                                        {{ userRequestTotals.onDemand }}
-                                    </div>
-                                    <div class="number-label">On Demand</div>
-                                </div>
-                            </div>
-                        </a>
-                    </div>
-                    <div class="col-md-2">
-                        <a class="big-number-link" href="{{appRoot}}/requests/active/WORKER/{{ deployUser }}" data-requests-active-filter="WORKER">
-                            <div class="well">
-                                <div class="big-number">
-                                    <div class="number" data-state-attribute="requests">
-                                        {{ userRequestTotals.worker }}
-                                    </div>
-                                    <div class="number-label">Worker</div>
-                                </div>
-                            </div>
-                        </a>
-                    </div>
-                    <div class="col-md-2">
-                        <a class="big-number-link" href="{{appRoot}}/requests/active/SCHEDULED/{{ deployUser }}" data-requests-active-filter="SCHEDULED">
-                            <div class="well">
-                                <div class="big-number">
-                                    <div class="number" data-state-attribute="requests">
-                                        {{ userRequestTotals.scheduled }}
-                                    </div>
-                                    <div class="number-label">Scheduled</div>
-                                </div>
-                            </div>
-                        </a>
-                    </div>
-                    <div class="col-md-2">
-                        <a class="big-number-link" href="{{appRoot}}/requests/active/RUN_ONCE/{{ deployUser }}" data-requests-active-filter="RUN_ONCE">
-                            <div class="well">
-                                <div class="big-number">
-                                    <div class="number" data-state-attribute="requests">
-                                        {{ userRequestTotals.runOnce }}
-                                    </div>
-                                    <div class="number-label">Run Once</div>
-                                </div>
-                            </div>
-                        </a>
-                    </div>
-                    <div class="col-md-2">
-                        <a class="big-number-link" href="{{appRoot}}/requests/active/SERVICE/{{ deployUser }}" data-requests-active-filter="SERVICE">
-                            <div class="well">
-                                <div class="big-number">
-                                    <div class="number" data-state-attribute="requests">
-                                        {{ userRequestTotals.service }}
-                                    </div>
-                                    <div class="number-label">Service</div>
-                                </div>
-                            </div>
-                        </a>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col-md-12">
-                <div class="page-header">
-                    <h2>My paused requests</h2>
-                </div>
-                {{#with pausedRequests}}
-                {{> requestsPausedBody}}
-                {{/with}}
-            </div>
-        </div>
-    {{/if}}
-    <div class="row">
-        <div class="col-md-12 table-staged">
-            <div class="page-header">
-                <h2>Starred requests</h2>
-            </div>
-            {{> requestsBody }}
-        </div>
-    </div>
-{{else}}
-    <div class="page-loader centered cushy"></div>
-{{/if}}
