--- SingularityUI/app/templates/dashboardTable/dashboardStarred.hbs
@@ -1,53 +1,0 @@
-{{#if haveStarredRequests}}
-    <table class="table table-striped sortable-theme-bootstrap" data-sortable>
-        <thead>
-            <tr>
-                <th>{{! Star column }}</th>
-                <th>Request</th>
-                <th>Type</th>
-                <th class="hidden-xs">Time of Last Deploy</th>
-                <th>Status</th>
-                <th class="visible-lg visible-md">Deploy user</th>
-                <th class="visible-lg visible-md">Instances</th>
-            </tr>
-        </thead>
-        <tbody>
-{{/if}}
-    {{#each starredRequests}}
-        <tr data-request-id="{{ request.id }}">
-            <td>
-                <a class="star" data-action="unstar" data-starred="true">
-                    <span class="glyphicon glyphicon-star"></span>
-                </a>
-            </td>
-            <td>
-                <a href="{{appRoot}}/request/{{ request.id }}">
-                    {{ request.id }}
-                </a>
-            </td>
-            <td>
-                {{humanizeText request.requestType}}
-            </td>
-            <td class="hidden-xs" data-value="{{ requestDeployState.activeDeploy.timestamp }}">
-                <span title="{{ requestDeployState.activeDeploy.timestamp }}">
-                    {{timestampFromNow requestDeployState.activeDeploy.timestamp}}
-                </span>
-            </td>
-            <td>
-                {{humanizeText state}}
-            </td>
-            <td class="visible-lg visible-md">
-                {{usernameFromEmail requestDeployState.activeDeploy.user}}
-            </td>
-            <td class="visible-lg visible-md">
-                {{ request.instances }}
-            </td>
-        </tr>
-    {{/each}}
-
-    {{#if haveStarredRequests}}
-    </tbody>
-</table>
-{{else}}
-<div class="empty-table-message"><p>No starred requests</p></div>
-{{/if}}
